+++ kolibri/plugins/coach/assets/src/views/common/commonCoachStrings.js
@@ -3,6 +3,8 @@ import { createTranslator } from 'kolibri.utils.i18n';
 const coachStrings = createTranslator('CommonCoachStrings', {
   // actions
   cancelAction: 'Cancel',
+  closeAction: 'Close',
+  continueAction: 'Continue',
   copyAction: 'Copy',
   createAction: 'Create',
   deleteAction: 'Delete',
@@ -15,6 +17,7 @@ const coachStrings = createTranslator('CommonCoachStrings', {
   newQuizAction: 'New quiz',
   previewAction: 'Preview',
   saveAction: 'Save',
+  saveChangesAction: 'Save changes',
   renameAction: 'Rename',
   showAction: 'Show',
   showMoreAction: 'Show more',
@@ -39,6 +42,7 @@ const coachStrings = createTranslator('CommonCoachStrings', {
   coachLabel: 'Coach',
   coachesLabel: 'Coaches',
   descriptionLabel: 'Description',
+  detailsLabel: 'Details',
   difficultQuestionsLabel: 'Difficult questions',
   entireClassLabel: 'Entire class',
   exercisesCompletedLabel: 'Exercises completed',
@@ -51,6 +55,8 @@ const coachStrings = createTranslator('CommonCoachStrings', {
   learnerLabel: 'Learner',
   learnersLabel: 'Learners',
   lessonLabel: 'Lesson',
+  lessonActiveLabel: 'Active',
+  lessonInactiveLabel: 'Inactive',
   lessonsLabel: 'Lessons',
   lessonsAssignedLabel: 'Lessons assigned',
   lessonsCompletedLabel: 'Lessons completed',
@@ -75,6 +81,8 @@ const coachStrings = createTranslator('CommonCoachStrings', {
   questionsCorrectLabel: 'Questions correct',
   questionsLabel: 'Questions',
   quizLabel: 'Quiz',
+  quizActiveLabel: 'Active',
+  quizInactiveLabel: 'Inactive',
   quizScoreLabel: 'Quiz score',
   quizzesLabel: 'Quizzes',
   quizzesAssignedLabel: 'Quizzes assigned',
@@ -133,6 +141,10 @@ const coachStrings = createTranslator('CommonCoachStrings', {
   percentage: '{value, number, percent}',
   ratio: '{value, number, integer} out of {total, number, integer}',
   ratioShort: '{value, number, integer} of {total, number, integer}',
+
+  // Errors
+  quizDuplicateTitleError: 'A quiz with that name already exists',
+  lessonDuplicateTitleError: 'A lesson with this name already exists',
 });
 
 const coachStringsMixin = {