+++ kolibri/plugins/coach/assets/src/views/plan/LessonSummaryPage/ManageLessonModals.vue
@@ -1,50 +1,22 @@
 <template>
 
   <div>
-    <AssignmentChangeStatusModal
-      v-if="lessonsModalSet === AssignmentActions.CHANGE_STATUS"
-      :modalTitle="$tr('changeLessonStatusTitle')"
-      :modalDescription="$tr('changeLessonStatusDescription')"
-      :modalActiveOption="$tr('changeLessonStatusActive')"
-      :modalInactiveOption="$tr('changeLessonStatusInactive')"
-      :active="currentLesson.is_active"
-      @changeStatus="handleChangeStatus"
-      @cancel="setLessonsModal(null)"
-    />
-
-    <AssignmentDetailsModal
-      v-else-if="lessonsModalSet === AssignmentActions.EDIT_DETAILS"
-      ref="detailsModal"
-      :modalTitle="$tr('editLessonDetails')"
-      :submitErrorMessage="$tr('saveLessonError')"
-      :modalTitleErrorMessage="lessonDetailEditorStrings.$tr('duplicateTitle')"
-      :initialDescription="currentLesson.description"
-      :showDescriptionField="true"
-      :isInEditMode="true"
-      :initialTitle="currentLesson.title"
-      :initialSelectedCollectionIds="initialSelectedCollectionIds"
-      :classId="classId"
-      :groups="learnerGroups"
-      @save="handleDetailsModalSave"
-      @cancel="setLessonsModal(null)"
-    />
-
     <AssignmentCopyModal
-      v-else-if="lessonsModalSet === AssignmentActions.COPY"
+      v-if="currentAction === AssignmentActions.COPY"
       :modalTitle="$tr('copyLessonTitle')"
       :assignmentQuestion="$tr('assignmentQuestion')"
       :classId="classId"
       :classList="classList"
-      @copy="handleCopy"
-      @cancel="setLessonsModal(null)"
+      @submit="handleSubmitCopy"
+      @cancel="closeModal"
     />
 
     <AssignmentDeleteModal
-      v-else-if="lessonsModalSet === AssignmentActions.DELETE"
+      v-if="currentAction === AssignmentActions.DELETE"
       :modalTitle="$tr('deleteLessonTitle')"
       :modalDescription="$tr('deleteLessonConfirmation', { title: currentLesson.title })"
-      @delete="deleteLesson({ lessonId: currentLesson.id, classId })"
-      @cancel="setLessonsModal(null)"
+      @submit="handleSubmitDelete"
+      @cancel="closeModal"
     />
   </div>
 
@@ -54,98 +26,105 @@
 <script>
 
   import find from 'lodash/find';
-  import { mapState, mapActions } from 'vuex';
-  import { crossComponentTranslator } from 'kolibri.utils.i18n';
   import CatchErrors from 'kolibri.utils.CatchErrors';
   import { ERROR_CONSTANTS } from 'kolibri.coreVue.vuex.constants';
-  import AssignmentChangeStatusModal from '../../plan/assignments/AssignmentChangeStatusModal';
-  import AssignmentDetailsModal from '../../plan/assignments/AssignmentDetailsModal';
+  import { mapState } from 'vuex';
+  import { LessonResource } from 'kolibri.resources';
   import AssignmentCopyModal from '../../plan/assignments/AssignmentCopyModal';
   import AssignmentDeleteModal from '../../plan/assignments/AssignmentDeleteModal';
   import { AssignmentActions } from '../../../constants/assignmentsConstants';
-  import LessonDetailEditor from '../../common/LessonDetailEditor';
-
-  const lessonDetailEditorStrings = crossComponentTranslator(LessonDetailEditor);
+  import { coachStringsMixin } from '../../common/commonCoachStrings';
 
   export default {
     name: 'ManageLessonModals',
     components: {
-      AssignmentChangeStatusModal,
-      AssignmentDetailsModal,
       AssignmentCopyModal,
       AssignmentDeleteModal,
     },
+    mixins: [coachStringsMixin],
+    props: {
+      // Should be 'COPY' or 'DELETE'
+      currentAction: {
+        type: String,
+      },
+    },
     computed: {
       ...mapState(['classList']),
-      ...mapState('classSummary', { classId: 'id' }),
-      ...mapState('lessonSummary', ['currentLesson', 'lessonsModalSet', 'learnerGroups']),
+      ...mapState('lessonSummary', ['currentLesson']),
       AssignmentActions() {
         return AssignmentActions;
       },
-      initialSelectedCollectionIds() {
-        return this.currentLesson.lesson_assignments.map(recipient => recipient.collection);
-      },
-      lessonDetailEditorStrings() {
-        return lessonDetailEditorStrings;
+      classId() {
+        return this.$route.params.classId;
       },
     },
     methods: {
-      ...mapActions('lessonSummary', [
-        'setLessonsModal',
-        'updateLessonStatus',
-        'deleteLesson',
-        'copyLesson',
-        'updateLesson',
-      ]),
-      handleChangeStatus(isActive) {
-        this.updateLessonStatus({ lessonId: this.currentLesson.id, isActive });
+      closeModal() {
+        this.$emit('cancel');
       },
-      handleCopy(selectedClassroomId, selectedCollectionIds) {
-        const payload = {
-          title: this.$tr('copyOfLesson', { lessonTitle: this.currentLesson.title }).substring(
-            0,
-            50
-          ),
-          description: this.currentLesson.description,
-          resources: this.currentLesson.resources,
-          collection: selectedClassroomId,
-          lesson_assignments: selectedCollectionIds.map(id => ({ collection: id })),
-        };
+      handleSubmitCopy(selectedClassroomId, selectedCollectionIds) {
+        const title = this.$tr('copyOfLesson', { lessonTitle: this.currentLesson.title }).substring(
+          0,
+          50
+        );
         const classroomName = find(this.classList, { id: selectedClassroomId }).name;
-        this.copyLesson({ payload, classroomName });
-      },
-      handleDetailsModalSave(payload) {
-        this.updateLesson({
-          lessonId: this.currentLesson.id,
-          payload: {
-            ...payload,
-            lesson_assignments: payload.assignments,
+        LessonResource.saveModel({
+          data: {
+            title,
+            description: this.currentLesson.description,
+            resources: this.currentLesson.resources,
+            collection: selectedClassroomId,
+            lesson_assignments: selectedCollectionIds.map(id => ({ collection: id })),
           },
         })
-          .then()
-          .catch(error => {
-            const errors = CatchErrors(error, [ERROR_CONSTANTS.UNIQUE]);
-            if (errors) {
-              this.$refs.detailsModal.handleSubmitTitleFailure();
-            } else {
-              this.$refs.detailsModal.handleSubmitFailure();
+          .then(() => {
+            // If copied to the same classroom, update the class summary
+            if (this.classId === selectedClassroomId) {
+              this.$store.dispatch('classSummary/refreshClassSummary');
             }
+            this.closeModal();
+            this.$store.dispatch('createSnackbar', this.$tr('copiedLessonTo', { classroomName }));
+          })
+          .catch(error => {
+            const caughtErrors = CatchErrors(error, [ERROR_CONSTANTS.UNIQUE]);
+            if (caughtErrors) {
+              this.$store.commit('CORE_CREATE_SNACKBAR', {
+                text: this.$tr('uniqueTitleError', {
+                  title,
+                  className: classroomName,
+                }),
+                autoDismiss: false,
+                actionText: this.coachStrings.$tr('closeAction'),
+                actionCallback: () => this.$store.commit('CORE_CLEAR_SNACKBAR'),
+              });
+            } else {
+              this.$store.dispatch('handleApiError', error);
+            }
+            this.closeModal();
+          });
+      },
+      handleSubmitDelete() {
+        const { title, id } = this.currentLesson;
+        return LessonResource.deleteModel({ id })
+          .then(() => {
+            this.$router.replace(this.$router.getRoute('PLAN_LESSONS_ROOT'), () => {
+              this.$store.dispatch('createSnackbar', this.$tr('lessonDeleted', { title }));
+            });
+          })
+          .catch(error => {
+            this.$store.dispatch('handleApiError', error);
           });
       },
     },
     $trs: {
-      changeLessonStatusTitle: 'Change lesson status',
-      changeLessonStatusDescription: 'Learners can only see active lessons',
       copyLessonTitle: 'Copy lesson to',
-      changeLessonStatusActive: 'Active',
-      changeLessonStatusInactive: 'Inactive',
       assignmentQuestion: 'Assign lesson to',
       deleteLessonTitle: 'Delete lesson',
       deleteLessonConfirmation: "Are you sure you want to delete '{ title }'?",
-      editLessonDetails: 'Edit lesson details',
-      newLesson: 'Create new lesson',
-      saveLessonError: 'There was a problem saving this lesson',
       copyOfLesson: 'Copy of { lessonTitle }',
+      copiedLessonTo: `Copied lesson to '{classroomName}'`,
+      uniqueTitleError: `A lesson titled '{title}' already exists in '{className}'`,
+      lessonDeleted: `Lesson '{title}' was deleted`,
     },
   };