+++ packages/kolibri-tools/.eslintrc.js
@@ -155,5 +155,6 @@ module.exports = {
     'kolibri/vue-no-unused-properties': ERROR,
     'kolibri/vue-no-unused-vuex-properties': ERROR,
     'kolibri/vue-no-unused-methods': ERROR,
+    'kolibri/vue-watch-no-string': ERROR,
   },
 };